Rewrite ActionState.
Sometimes you have this idea in your head of how things should go, and
then you try to translate it into code, and then the language doesn't
support what you were thinking, and then it ends up awful - but you
don't realize because you see the idea in your head when you look at
the code.

This is one of those times.