Test and fix current word search commands
Add tests for `g*` and `g#` by themselves and in combination with `n` (to step to next command).

Previously the `#` and `*` commands (and, by virtue of having been copied, the `g#` and `g*` commands introduced in a previous commit) would immediately search and filter any matches that were not full word matches.

ie. they would skip matches "hello" when searching for "he".

This would work fine for `*` or `#` directly, but subsequent uses of `n` or `N` would match incorrectly since the word bounds were not included in the search state.