README enhancements (#1547)
* Reformat the heading (add logo, centered styling)

The HTML is supported in the VS Code Marketplace site too, and this is
layout is shamelessly stolen from the Chrome Debugger plugin:
https://marketplace.visualstudio.com/items?itemName=msjsdiag.debugger-for-chrome

* Normalize how we reference the real Vim plugins

This what their repos are named, and make our README look more
consistent.

* Remove reference to useCtrlKeys in features list

It's enabled by default, and so isn't really a feature.

* Bulk refactoring of README

* Re-arrange certain sections based on importance
* Promoted plugin info to it's own section
* Add contents
* Bunch of other small formatting tweaks

* Simply plugins point in features

They're outlined nicely in the contents area just below, seems like
overkill to have them all listed here too.

* Add Contents link to status bar customization (vim-airline)

* Push FAQ further down the page

The tricks list is cooler (read: more important)

* Additional language/linking tweaks

* Add 'json' language to code blocks

And clean up extra closing backticks

* Prefer "setting" over "option" or "configuration"

This is in-line with terminology in VSCode (i.e. `settings.json`).

* Fix heading numbers and use Sentence case

* Fix formatting and indentation issues

* Move Key remppaing under settings

And move `useCtrlKeys` and `handleKeys` into VSCodeVim Settings.

* Add OS-specific setup sub-sections

* More heading tweaks

* Add more links to useCtrlKeys and handleKeys

* Fix formatting with backticks

The stuff between the <...> was getting drop because Markdown parsers
think that it's HMTL tags ;)

* Fix typo

* More heading level and case fixes

* Add quick-link to `gc` docs

* Format VSCode Vim tricks list consistently

* Test: backticks around settings headings

* Make setting's headings map to JSON keys

* Fix Contents link to Emulated plugins

* Tweaks to Contributing and Release notes

* Tight up features some more

* Introduce "Getting started" section with OS-setup info

It didn't make sense to have Settings-related info here. The quick
example settings has been more to the start of the Settings section.

* Add sub-section on compatibility with link to ROADMAP

* Remove duplicate info about key repeat issues on Mac

* Tweak waffling

* Small grammar fix