Join HTML on single line to prevent extraneous <br>s
It would appear that the marketplace website (https://marketplace.visualstudio.com/items?itemName=vscodevim.vim)
was converting line-breaks in the HTML source into <br> tags, creating heaps of ugly whitespace.
This is a test to see if we remove most of the line-breaks that it'll clean all the extra whitespace.