Fix extension config access & overwrite
In order to properly access and overwrite the extension configuration
values, the name of the private field needed to match the value found in
the configuration file, while the getter & setter were given a different
name. Also, to properly use the vscode.WorkspaceConfiguration.update()
API, ConfigurationTarget.Global had to be specified. The value true
could've also been used, but the enum is more explicit.