Fix vertical split tests
The vertical split tests were formerly polling the window, waiting for
the number of text editors to change. This is now triggered based on the
vscode.window.onDidChangeVisibleTextEditors event.

Also, a setup/teardown bug was fixed so the setup and teardown is
performed for each test, not once for the suite.