Added separate condition for reverse select
Given the reverse selection only selected the first character but forward selection worked fine, we change a reverse selection into a forward selection at the background so that everything works fine.
Possible cause of the bug: getText() probably started from the right of the stop and then moved backward until it find's start but if it cannot do so, it just selects the first character as can be seen in the case of backward selection.