Prevent error on loading search history if no active editor on startup
We were seeing the error on startup with globalState.loadSearchHistory().

TypeError: Cannot read property 'document' of undefined

This happened if the editor loaded with no active tab when initializing VSCodeVim.

When creating a new SearchState in loadSearchHistory, it was calling searchState.set searchString, which calls _recalculateSearchRanges, which tried to access the document that didn't exist.