Reset the command list in ModeHandler.runAction if in insert mode
Line 382 was checking if there was an action ran AND if the current
state was not insert mode. This caused a bug which left the commandList
with a lingering `i` which broke remapped bindings for the following
key. Remove this check so that the command list is clear upon entering
insert mode.