Fix `C-n` and `C-p` in autocomplete for multi cursor (#3283)
This fixes jumping the autocomplete options in multi cursors so that it
only jumps one at a time. Without this fix it jumps once _per cursor_,
so if I have 3 cursors, I am forced to traverse the list 3 at a time.

I initially found this as part of https://github.com/VSCodeVim/Vim/issues/3282.