Put'ing text ending in newline should behave like LineWise:
Given the block:

      1234567
    1.def foo
    2.  a
    3.  b
    4.end

Putting the cursor inside and running `2yy` to copy two lines will put
"a\n  b" into the register and set LineWise mode.

Alternatively with the cursor on r2c1, pressing `vj$` will fill the
buffer with `  a\n  b\n` in CharacterWise mode.

In vim, Put'ing either of these has the same effect. That is, given:

      1234567
    1.// word
    2.// etc.

Putting the cursor at r1c1 and running `p` results, in *either* case,
in:

      1234567
    1.// word
    2.  a
    3.  b
    4.// etc.

In VSCode, however, the latter case -- the CharacterWise with a trailing
newline -- we get:

      1234567
    1./  a
    2.  b
    3./ word
    4.// etc.

Vim seems to use the presence of a trailing newline as an indication of
whether a Put should be considered LineWise, as does this PR.