Fix typo in README.md
In the README, a number of nifty tricks are referenced, however, the text incorrectly refers to these features as being Vim features. While `gd` and `gq` are standard, `af`, `gb` and `gh` are not standard vim features. What is probably indented is that these features are VSCode features adapted for VSCodeVim.