Respect `editor.autoClosingBrackets` and `editor.autoClosingQuotes` when deleting a bracket/quote (#3941)
When we delete a "pairable" character like (, we look to the right to see if there's the matching ) and delete it is well. We should not do that if auto-matching is disabled in settings. This behavior regressed because the type of editor.autoClosingBrackets was changed from a bool to a more fine-grained enum. There's also editor.autoClosingQuotes now, so we should respect that too.
Fixes #3936