Fix key mapping for <C-n> and <C-p>
Fixes #4011
VSCode recently introduced the `||` operator in `when` expressions, which we used to combine several separate definitions.
`||` has lower precedence than `&&`, but we needed `||` to bind more tightly. That's what parentheses are for, right?
Parentheses are apparently unsupported by VSCode, so we need to write a much longer, more repetitive `when` expression.