Refactor and fix edge cases of <C-u> and <C-d> (#4117)
Our implementation of <C-u> and <C-d> has never quite been right - currently, <C-d> gets stuck at the bottom and <C-u> jumps to the top as soon as it's visible. We also stopped paying any attention to the vim.scroll option at some point.

This commit fixes all of the above and simplifies the logic.