Fix numeric string parse bug caused by stateful regexes (#4221)
When incrementing or decrementing a number, the NumericString.parse class is used to parse strings in to numbers of various formats. However, it used global (stateful) regexes to do this, which meant that funky things could happen after the first call to the method. This PR fixes the issue by simply removing the g flag from some regexes.

This fixes issues, for example, where pressing <C-a> on certain numbers more than once lead to numeric prefixes getting stripped or other odd behavior.

Fixes #3545