Fix cursor off after leaving multi-cursor mode
Previously when leaving multi-cursor mode with <Esc> the cursor would
move one column to the left. For instance, when triggering multi-cursor
mode on a single word, the cursor would end at the second to last
character. While this is expected behavior when leaving insert mode, it
is not for visual mode (which multi-cursor mode is based on). To fix,
remove special case handling altogether.

Fixes #3775.