Put the recorded state (keys you've hit) in a new StatusBarItem which drifts right
We could put this behind a setting, I suppose, but I'll wait for someone to complain first, since this is how vim behaves.
I've set the StatusBarItem priority to Number.MAX_SAFE_INTEGER (and MIN_SAFE_INTEGER for the other), which is a bit obnoxious, but I think we want them to be hugging the center.
Fixes #4006