Elevate `doesntChangeDesiredColumn` to BaseAction (#4424)
Fixes #4296

Several actions, such as `zz` and `<C-e>` changed the "desired column" erroneously.
This was most noticeable if you did something like the following:
`$` (which sets the desired column to EOL), then go to a shorter line and `zz` to center the viewport, then back to the longer line. We want to end at the end of the longer line, which is now done correctly.