Refactor `Position` and `PositionDiff`
The biggest change is in how PositionDiffs are constructed - now with a object parameter, so it's much easier to understand what the intention is.
BOL diffs no longer accept a character - they bring you to the beginning of the line.
And a few other small tweaks. This probably should've been split across a couple commits, but too late for that now.