Move commands that only pertain to CommandLine/Search modes to a separate file
I'm isolating these mostly to cut down on the massive size of `actions.ts`, but also because I want to combine many of the actions here so that we don't have so much duplicated code.
Also, `CommandInsertIn___Mode` is really a whole bunch of commands in one, and splitting them up in a separate file will be less messy than contributing even more to the bloat in `actions.ts`.