Implement  _isDocumentTextNeeded() (#4386)
For reference, see #4380 and #3920 .

As mentioned in #4380 HistoryTracker.addChange() performance gets worse the larger the file and the faster we make individual changes, because on each change we would get the entire document's text. We don't use the individual changes, so we really only care about the initial step and the final one.

A new historyStep is created when we first enter a new mode and finalized after we exit it. We can keep track of the currentMode internally, and when it is different to vimState.currentMode we know we have switched. This way we prevent making unnecessary calls to TextDocument.getText() while changes are still in progress.

This commit also activates the optimization from #4451 when `experimentalOptimizations` is enabled.