Fix cursor movement with large folded regions.
With a folded region larger than 15 lines, when the cursor is below the
folded region, it is always 15 lines below the end of the first region,
so the existing logic will incorrectly tell vim to always re-center the
cursor. Instead, we should check if the cursor is either below all
visible regions, or above all visible regions.