After a failed search, don't swallow the next key strok
When a search failed, we were throwing a VimError which was caught by ModeHandler and displayed in the status bar. That's well and good, but the disruption of the control flow was preventing the "keys pressed" array from being cleared. Rather than trying to figure out exactly what was causing this, I decided to simply maintain our usual control flow by reporting the error and returning from the action. Let's leave exceptions to exceptional cases.
Fixes #4658