Fix error after `"_cw<left>`
For some reason, running an operator like `cw` with a register will cause `recordedState.operator` to be undefined when using arrow keys immediately afterward. The easiest way of fixing this was to add some type safety (operator can be undefined, so let's tell typescript that) and check it before utilizing it for `mightChangeDocument`.
Fixes #4640