Various small refactors in ModeHandler
Let's face it - ModeHandler is pretty incomprehensible.
This is partly inherent complexity (there are a lot of weird edge cases in vim and in VSCode, and they often interact to form weird corner cases).
This commit is an attempt to reduce the incidental complexity caused by poorly named variables, unnecessarily convoluted logic, etc. so that we can more effectively deal with the inherent complexity.
No behavior changes here.