Refactor: begin pulling code out of Position and Range
The motivation for this is a desire to eventually replace our Range with a Cursor class that carries with it some information useful for multi-cursor mode (#4585).
That's potentially a big lift, so I want to start by simplifying as much as possible - most of Position.ts belongs somewhere else, anyway.