Fix behaviour when pasting lines over selection (#3891)
Currently when pasting lines of text (i.e. text in register is linewise)
over an existing selection, the code incorrectly yanks the selected
lines of text into the register specified by the command. This is
incorrect behaviour as the selected text should always be yanked into
the unnamed register (or clipboard register if appropriate).

Fixes #2895, fixes #2939.