Compress DocumentContentChangeAction changes in the basic case
Rather than inserting characters one by one when repeating an insertion via dot or a macro, we now do it all in one go.
There are cases that this doesn't cover (most notably backspace), but this should be a safe change that will improve most cases of repeated insertions drastically.

Refs #4809