Fix: cmd+D now jumps and scrolls window to the last selected word (#4725)
There is an issue in the current version, where if you search and select with cmd+D, if the next match is out of view of the current window, the window is flashed to that next match but then it immediately returns to the first result. This PR fixes this issue.

The problem was that while the action inserts multiple cursos into the workspace, the updateView() function always tracks the first one (there are also 2 functions which call the updateView() right after each other: handleKeyEventHelper() and handleSelectionChange(), but I got scared of touching the latter function just by reading the comment above it, so I tinkered with updateView()).

I'm aware that this rather a patch than a true fix and is utterly inelegant (it checks whether the last action was cmd+D and reacts accordingly - maybe if other actions are found, which would benefit from this solution, we can expand this condition), so I'm open to suggestions. The flipside of this is that the patch is isolated only for this action, so it shouldn't break any previously correct behaviour.

Fixes #880