Fix recording macro not showing after 'setText' (#4878)
The Recording @ message was being erased because every time the StatusBar.setText is called it replaces the existing text with the new message.

This happens, for example, when the :write command calls setText or when you undo something, which might happen in the middle of the macro recording and erases the recording message.

The message would only update again with the Recording @ after a mode change. This makes it update after the next key press, which means if you undo you get the undo message until you press some key which replaces that message with the clear text that includes the macro recording.

This is the Vim behavior but I think Neovim always shows it. We could make it always show up by appending it to text on the StatusBar.setText(text) function as well. Some users might prefer it that way.

Fixes #4753