Fixed the workaround on line 713 so that it actually works
The workaround for linux key changes & vsCode, which is displayed on line 713 ( "keyboard.dispatch": "keyCode"  ) had 'keycode' but it should be 'keyCode'. Using 'keycode' does not resolve the issue. The link 'known issue in VS Code' has the solution in it, and they also use keyCode. I've tested that keyCode resolves the issue still.