Fix 'v' enter visual mode not clearing commandList (#5000)
When going to visual mode with key 'v' since it is not a complete action the modeHandler wasn't resetting the recordedState.commandList which meant that if the next key came before timeout had passed it would add that key to the commandList that still had the 'v'. Then it would send that list to the remappers.

If you had a remap in visual mode like L -> $ after pressing vL it would send ["v", "L"] to the remappers so it wouldn't find any remap. Only after timeout ended would it send only the last key of commandList and in that case the remappers would find the remap.

Fixes #4991