Textobject argument improvements (#5004)
This PR improves the argument movement, by handling some more edge-cases and being less aggressive on the existing whitespaces in the argument (i.e. preserving indentation-level) when doing the inner argument movement.

One of the edge-cases was not checking that the cursor is actually inside the delimiters, i.e. that we actually are inside an argument list. This caused the movement to span to the neighboring separators (,) across the buffer. This PR adds a check to fail the movement right at the beginning.
Unfortunately this requires 2 scans for delimiters/separators on successful movements. I think this is okay for the current scope of argument lists, since the search range is quite limited on the second scan. In a future update one could rework the logic, so that the find-functions actually keep track of all the visited separators on the first scan.

On another note: This early failure could be relaxed in a further update, so that comma-separated lists when they are on a single line are a valid movement (there may be several cases, where this feature could be useful, e.g. from x import a, b, c in python).

This is a direct successor to #4653