Create placeholders in subparser map for not-yet-implemented commands
This gives an error message which makes it clear that we understood what you were trying to do, but it's not implemented yet.
Also makes it easier to see, as a maintainer, what is left to be done in terms of ex commands.
Not a complete list, but commonly requested commands are present.