Surround refactor
Surround is one of the ugliest ducklings in the codebase; here's an attempt to make it a bit saner.
None of the structure has really changed (yet); most of this is just rearrangement, clean up, comments, etc.
Only real behavioral change is a reduction of the number of transformations done from 4 to 2 when deleting or changing a surrounding tag. I'm not sure why it was written as it was, and it still seems to work 🤞