Added "Improve Existing Issues" notes
For the "duplicate" link, I provided a sample search for "cursor" which returns 200+ issues. Many of which probably aren't duplicates, but I thought that might be a good starting point. MIght be better to link to a blank search for open issues.

For the "old" link, I gave page 25. Currently there are 34 pages, so I thought that would be sufficient insulation to still be "old" but not get shwooped off later

I saw a few folks claim that VSCodeVIM Team gives precedence to issues with more thumbs up, so I included that recommendation. I didn't want to say that so explicitly though, since I don't actually know.