Introducing: the Transformer
This is (currently) very thin layer around the Transformation list (previously held directly by RecordedState).
My goal in adding this is to consolidate the logic around Transformations into one place, where we can more easily audit it (with logging, for instance), and make it smarter at some point down the road.
In the medium term, this should be made the single boundary between Actions and the Document they modify.