Small VSCodeContext refactor with performance enhancement
Ah, my nemesis falsiness strikes again. Due to a small oversight, setting a context value to `false` when it already was `false` triggered an unnecessary call to `setContext` (which is surprisingly expensive).
This is mostly felt when changing config when `handleKeys` has any `false` values in it. This is frequent if you use a setting like `statusBarColorControl` or `smartRelativeLine.
In addition, I made some changes to capitalization and made `VSCodeContext` a static class.
Refs #3868, I suppose