Use `yarn` instead of `npm`
We've had a yarn.lock alongside package-lock.json for a while now, and there should really only be one.
Yarn is faster and more deterministic, and while it's one more download, it's almost entirely a drop-in replacement for npm.
There is Yarn 2.x (also known as Berry for some reason), but from what I've read, it requires a more substantial migration and is more controversial. Let's stay on 1.x for now, and I'll look into upgrading at some point down the line.