Attempt to untangle the put logic somewhat
This is easily one of the worst areas of the codebase.
It really needs a fundamental refactor (closer to a rewrite, really), but a good first step is to make it a bit more understandable as-is.
This attempts to change no behavior or even fundamental logic, just to make it a bit easier to reason about.