Re-write put from scratch
I finally bit the bullet, deleted all of our put logic, and re-implemented it from scratch.
This includes `p`, `P`, `gp`, `gP`, `]p`, `[`, and `:put`.

In addition to cleaner, more maintainable code, this resulted in:
  - Using replace transformations, which means better performance
  - Fixing VisualBlock put and block-wise put (fixes #1796)
  - Fixing put when the register contains a macro (fixes #3714)
  - Fixing various other corner cases, particularly with regards to cursor placement
  - Beefing up our test suite

I'm sure there are still obscure corner cases that we handle wrong, but it'll be easier going forward to fix them.