Fix logic for merging content changes that came from a tab completion
This makes repeating an insertion that contains a tab completion work again (this broke when we moved to using transformations). As a bonus, it's all done in a single transformation. Yay!
The core logic here is essentially identical to that in HistoryTracker, and it should probably be abstracted out and unit tested.
This deserves integration testing as well, but that'll require a bit of nuance, and it's getting late.
Fixes #6610