Show command and search when mode name is disabled (#7021)
When `showmodename` is disabled the mode is hidden but the search and
in-progress command text is also hidden.

This change updates the logic to take the command line and search
progress into account when determining if the status should be updated.

The old behavior of hiding all the text can be achieved by right
clicking the status bar and hiding the "Vim Command Line" entry. The
only difference is that macro information will also be hidden.