Refactor `BaseAction.mustBeFirstKey` out
An operator must be followed by a count, an operator (`dd`) or with a motion-like action.
This was, as far as I can tell, the use of `mustBeFirstKey`, so let's just encode that rule and drop the property.
Introducing `isNumber` for this feels a bit silly (I'll refactor later), but I think this is an improvement.