Completely re-write core of ex command parser
Here's a refactor that's long overdue: ex command parsing.
We have (for some reason) been using a custom-built parser which is
incomplete, buggy, and complicated.
Now we're using Parsimmon, a parser combinator library.

This fixes a few bugs and implements a few small features, but it's
mostly a small step toward more complete support for Vimscript.

Some next steps:
  - Support patterns in LineRanges
  - Improve ex command argument parsing by using Parsimmon
  - Support comments, combining commands with `|`