Fixed fadeout bug and actions bugs during SneakLabelMode
Added more lines to be faded out when using label mode to prevent the
user not seeing every line properly dimmed.

Added a way to escape label mode using the Escape key

Fixed the actions not properly being triggered when in labelMode and pressing a key that was not a label. For example pressing y in label mode (y is by default not in the label list) now properly trigger they key and wait for a motion.