Big refactor of command line & search
This removes a lot of code duplication, creates a better abstraction,
makes testing easier, and lays the groundwork for some features like:
- `:s` and similar commands working with `hlsearch` (#4934)
- The expression register (#4383)
- Better `incsearch` support (#4837)

The central change here is to take various pieces of global state which
were scattered and throughout the project and bring them together,
localizing them where possible to an object (`CommandLine`) held by
`VimState` (this naturally fixes #7038).

This applies to both ex commands and the search command line, which
means we finally have a uniform interface for them!