<C-g> and <C-t> in SearchInProgress mode
Keep track of current match during incsearch, allowing user to navigate between matches with <C-g> and <C-t>.

Current match depends on 3 things:
 - the search command count
 - the cursor's starting position
 - the number of times the user has pressed <C-g>/<C-t>
The first two are available through a vimState object.
We track the third with a field in the SearchCommandLine object, whose
new getCurrentMatchRange and getCurrentMatchPosition functions compute
information about the current match for a given vimState.

Also a minor change to backward search for consistency with vim:
When finding next match backwards, look for range whose start is strictly before the start pos,
instead of a range whose end is at or before the start pos.