Implement `inccommand` (#7416)
This PR introduces the option `inccommand`, which enables a live preview of `:s` commands as they are typed. It also makes some improvements to interactive substitution with the `c` flag. 

The feature is configured by setting `vim.inccommand` to `'append'` or `'replace'` (or `''` to disable).
When set to `'append'`, substitution previews are inserted after each highlighted match, a behavior found in some popular vim emulators like spacemacs.
When set to `'replace'`, the substitution previews replace each match, which is consistent with the default behavior of neovim.

The config keys `vim.substitutionColor` and `vim.substitutionTextColor` are provided to control the appearance of preview text. 

Fixes #6656, fixes #6655