Make vim.textwidth be configurable per-language
Currently the vim.textwidth setting is global which means that all
languages must share the same value. This is fine for a lot of cases but
for some cases like git commits it can be useful to have a custom value.

This updates setting configuration to be "language-overridable" which
allows VSCode to autocomplete the value in the settings page for
languages. This allows us to update the `textwidth` property to a getter
where VSCode will automatically return the correct value.

May close https://github.com/VSCodeVim/Vim/issues/4993