Fixes #9801. Save prevResult when it is a Position (#9802)
For motions with a count (e.g., 3[{), save the previous iteration's
result when that result is a Position.

With a motion like 99[{, eventually (in typical files), an iteration
will fail to find an unmatched { and therefore return prevResult.
However, <count>[{ returns results of type Position, and therefore,
prevResult was not being updated for it. This caused execActionWithCount
to return prevResult's initial value of failedMovement. This in turn led
to the cursor not moving at all, when it should have moved to the last
unmatched {, i.e., the Position of the last successful iteration.

This commit fixes this issue by making sure prevResult is updated when
result is a Position.